Refactor Admin components
