+++ app/javascript/components/admin/roles/RoleRow.jsx
@@ -4,7 +4,7 @@ import { Badge, Dropdown } from 'react-bootstrap';
 import { DotsVerticalIcon, PencilAltIcon, TrashIcon } from '@heroicons/react/outline';
 import { Link } from 'react-router-dom';
 import Modal from '../../shared_components/modals/Modal';
-import DeleteRoleForm from '../../forms/DeleteRoleForm';
+import DeleteRoleForm from './forms/DeleteRoleForm';
 
 export default function RoleRow({ role }) {
   return (
