+++ app/javascript/components/admin/roles/forms/CreateRoleForm.jsx
@@ -4,11 +4,11 @@ import {
   Button, Stack,
 } from 'react-bootstrap';
 import PropTypes from 'prop-types';
-import Form from './Form';
-import Spinner from '../shared_components/utilities/Spinner';
-import FormControl from './FormControl';
-import { createRoleFormConfig, createRoleFormFields } from '../../helpers/forms/CreateRoleFormHelpers';
-import useCreateRole from '../../hooks/mutations/roles/useCreateRole';
+import Form from '../../../shared_components/forms/Form';
+import Spinner from '../../../shared_components/utilities/Spinner';
+import FormControl from '../../../shared_components/forms/FormControl';
+import { createRoleFormConfig, createRoleFormFields } from '../../../../helpers/forms/CreateRoleFormHelpers';
+import useCreateRole from '../../../../hooks/mutations/roles/useCreateRole';
 
 export default function CreateRoleForm({ handleClose }) {
   const createRole = useCreateRole({ onSettled: handleClose });
