+++ app/javascript/components/admin/server_recordings/ServerRecordingRow.jsx
