+++ app/javascript/components/admin/server_recordings/ServerRecordings.jsx
@@ -3,13 +3,13 @@ import Card from 'react-bootstrap/Card';
 import {
   Col, Container, Row, Tab, Stack, Button,
 } from 'react-bootstrap';
-import AdminNavSideBar from './shared/AdminNavSideBar';
-import SearchBarQuery from '../shared_components/search/SearchBarQuery';
-import RecordingsList from '../recordings/RecordingsList';
-import useServerRecordings from '../../hooks/queries/admin/server_recordings/useServerRecordings';
-import ServerRecordingRow from './server-recordings/ServerRecordingRow';
-import useRecordingsReSync from '../../hooks/queries/recordings/useRecordingsReSync';
-import Pagy from '../shared_components/Pagy';
+import AdminNavSideBar from '../AdminNavSideBar';
+import SearchBarQuery from '../../shared_components/search/SearchBarQuery';
+import RecordingsList from '../../recordings/RecordingsList';
+import useServerRecordings from '../../../hooks/queries/admin/server_recordings/useServerRecordings';
+import ServerRecordingRow from './ServerRecordingRow';
+import useRecordingsReSync from '../../../hooks/queries/recordings/useRecordingsReSync';
+import Pagy from '../../shared_components/Pagy';
 
 export default function ServerRecordings() {
   const [input, setInput] = useState();
