+++ app/javascript/components/admin/site_settings/administration/Administration.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Row } from 'react-bootstrap';
-import LinksForm from '../../forms/admin/LinksForm';
-import useUpdateSiteSetting from '../../../hooks/mutations/admin/site_settings/useUpdateSiteSetting';
+import LinksForm from './LinksForm';
+import useUpdateSiteSetting from '../../../../hooks/mutations/admin/site_settings/useUpdateSiteSetting';
 
 export default function Administration({ terms, privacy }) {
   return (
