+++ app/javascript/components/admin/site_settings/registration/forms/RegistrationForm.jsx
@@ -4,10 +4,10 @@ import {
   Button, InputGroup,
 } from 'react-bootstrap';
 import { useForm } from 'react-hook-form';
-import Form from '../Form';
-import { RegistrationFormFields, RegistrationFormConfig } from '../../../helpers/forms/RegistrationFormHelpers';
-import Spinner from '../../shared_components/utilities/Spinner';
-import FormControlGeneric from '../FormControlGeneric';
+import Form from '../../../../shared_components/forms/Form';
+import { RegistrationFormFields, RegistrationFormConfig } from '../../../../../helpers/forms/RegistrationFormHelpers';
+import Spinner from '../../../../shared_components/utilities/Spinner';
+import FormControlGeneric from '../../../../shared_components/forms/FormControlGeneric';
 
 export default function RegistrationForm({ value, mutation: useUpdateSiteSettingsAPI }) {
   const updateSiteSettingsAPI = useUpdateSiteSettingsAPI();
