+++ app/javascript/components/recordings/forms/DeleteRecordingForm.jsx
@@ -4,7 +4,7 @@ import {
   Button, Stack,
 } from 'react-bootstrap';
 import PropTypes from 'prop-types';
-import Form from '../../forms/Form';
+import Form from '../../shared_components/forms/Form';
 import Spinner from '../../shared_components/utilities/Spinner';
 
 export default function DeleteRecordingForm({ mutation: useDeleteAPI, recordId, handleClose }) {
