+++ app/javascript/components/rooms/room/forms/DeleteRoomForm.jsx
@@ -4,7 +4,7 @@ import {
   Button, Stack,
 } from 'react-bootstrap';
 import PropTypes from 'prop-types';
-import Form from '../../../forms/Form';
+import Form from '../../../shared_components/forms/Form';
 import Spinner from '../../../shared_components/utilities/Spinner';
 
 export default function DeleteRoomForm({ mutation: useDeleteRoomAPI, handleClose }) {
