+++ app/javascript/components/rooms/room/presentation/forms/DeletePresentationForm.jsx
@@ -5,7 +5,7 @@ import {
 } from 'react-bootstrap';
 import { useParams } from 'react-router-dom';
 import PropTypes from 'prop-types';
-import Form from '../../../../forms/Form';
+import Form from '../../../../shared_components/forms/Form';
 import Spinner from '../../../../shared_components/utilities/Spinner';
 import useDeletePresentation from '../../../../../hooks/mutations/rooms/useDeletePresentation';
