+++ app/javascript/components/shared_components/forms/FormLogo.jsx
