+++ app/javascript/components/users/authentication/forms/SigninForm.jsx
@@ -9,8 +9,8 @@ import { useForm } from 'react-hook-form';
 import { Link } from 'react-router-dom';
 import HCaptcha from '@hcaptcha/react-hcaptcha';
 import { toast } from 'react-hot-toast';
-import FormControl from '../../../forms/FormControl';
-import Form from '../../../forms/Form';
+import FormControl from '../../../shared_components/forms/FormControl';
+import Form from '../../../shared_components/forms/Form';
 import { signinFormFields, signinFormConfig } from '../../../../helpers/forms/SigninFormHelpers';
 import Spinner from '../../../shared_components/utilities/Spinner';
 import useCreateSession from '../../../../hooks/mutations/sessions/useCreateSession';
