+++ app/javascript/components/users/authentication/forms/SignupForm.jsx
@@ -3,8 +3,8 @@ import { Button, Container, Stack } from 'react-bootstrap';
 import { useForm } from 'react-hook-form';
 import HCaptcha from '@hcaptcha/react-hcaptcha';
 import { toast } from 'react-hot-toast';
-import FormControl from '../../../forms/FormControl';
-import Form from '../../../forms/Form';
+import FormControl from '../../../shared_components/forms/FormControl';
+import Form from '../../../shared_components/forms/Form';
 import { signupFormConfig, signupFormFields } from '../../../../helpers/forms/SignupFormHelpers';
 import Spinner from '../../../shared_components/utilities/Spinner';
 import useCreateUser from '../../../../hooks/mutations/users/useCreateUser';
