+++ app/javascript/components/users/password_management/ResetPassword.jsx
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 import Card from 'react-bootstrap/Card';
 import { useParams } from 'react-router-dom';
 import useVerifyToken from '../../../hooks/mutations/users/useVerifyToken';
-import FormLogo from '../../forms/FormLogo';
+import FormLogo from '../../shared_components/forms/FormLogo';
 import ResetPwdForm from './forms/ResetPwdForm';
 import Spinner from '../../shared_components/utilities/Spinner';
