+++ app/javascript/components/users/password_management/forms/ForgetPwdForm.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Button, Stack } from 'react-bootstrap';
 import { useForm } from 'react-hook-form';
-import FormControl from '../../../forms/FormControl';
-import Form from '../../../forms/Form';
+import FormControl from '../../../shared_components/forms/FormControl';
+import Form from '../../../shared_components/forms/Form';
 import { forgetPwdFormConfig, forgetPwdFormFields } from '../../../../helpers/forms/ForgetPwdFormHelpers';
 import Spinner from '../../../shared_components/utilities/Spinner';
 import useCreateResetPwd from '../../../../hooks/mutations/users/useCreateResetPwd';
