+++ app/javascript/components/users/password_management/forms/ResetPwdForm.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Button, Stack } from 'react-bootstrap';
 import { useForm } from 'react-hook-form';
-import FormControl from '../../../forms/FormControl';
-import Form from '../../../forms/Form';
+import FormControl from '../../../shared_components/forms/FormControl';
+import Form from '../../../shared_components/forms/Form';
 import { resetPwdFormConfig, resetPwdFormFields } from '../../../../helpers/forms/ResetPwdFormHelpers';
 import Spinner from '../../../shared_components/utilities/Spinner';
 import useResetPwd from '../../../../hooks/mutations/users/useResetPwd';
