+++ app/javascript/components/users/user/forms/AvatarForm.jsx
@@ -3,8 +3,8 @@ import { useForm } from 'react-hook-form';
 import { Button } from 'react-bootstrap';
 import { yupResolver } from '@hookform/resolvers/yup';
 import PropTypes from 'prop-types';
-import FormControl from '../../../forms/FormControl';
-import Form from '../../../forms/Form';
+import FormControl from '../../../shared_components/forms/FormControl';
+import Form from '../../../shared_components/forms/Form';
 import useCreateAvatar from '../../../../hooks/mutations/users/useCreateAvatar';
 import { validationSchema, avatarFormFields } from '../../../../helpers/forms/AvatarFormHelpers';
